+++ src/directives/async-replace.ts
@@ -12,7 +12,7 @@
  * http://polymer.github.io/PATENTS.txt
  */
 
-import {directive, Directive, NodePart} from '../lit-html.js';
+import {directive, NodePart, Part} from '../lit-html.js';
 
 /**
  * A directive that renders the items of an async iterable[1], replacing
@@ -32,9 +32,12 @@ import {directive, Directive, NodePart} from '../lit-html.js';
  * @param mapper An optional function that maps from (value, index) to another
  *     value. Useful for generating templates for each item in the iterable.
  */
-export const asyncReplace =
-    <T>(value: AsyncIterable<T>, mapper?: (v: T, index?: number) => any):
-        Directive<NodePart> => directive(async (part: NodePart) => {
+export const asyncReplace = directive(
+    <T>(value: AsyncIterable<T>, mapper?: (v: T, index?: number) => any) =>
+        async (part: Part) => {
+          if (!(part instanceof NodePart)) {
+            throw new Error('asyncReplace can only be used in text bindings');
+          }
           // If we've already set up this particular iterable, we don't need
           // to do anything.
           if (value === part.value) {